Moving blp_positon util to tmutils.
